implements cloudformation controller interface
Previously implementing cloudformation-based services required:

* creating a kubernetes object spec-type (+ crd)
* creating a cloudformation template-type unique to the spec-type
* creating a controller-type unique to both the object-spec type and the
  template-type to tie them together and knows how to extract values in
  to secrets and deal with the reconciliation loops, making use of a
  shared reconciler-func where possible.

Problems with this structure:

* This required quite a bit of boilerplate code that was very similar
  between implementations of the tepmlate-types and the controller-types
* Very slight differences in the boilerplate controller code made it
  easy to miss differing behaviour
* The interface for the shared reconciler code was leaking it's
  implementation requiring the controllers to perform fairly complex
  handling of the responses.
* The controllers were difficult to test without mocking out the entire
  reconciler (which itself contained logic that should be tested)

To address these issues, with an aim of making the controller code more
testable, the following interfaces where created...

cloudformation.Stack:

Any type that implements cloudformation.Stack can produce a
cloudformation template of itself, ready to be consumed by the
cloudformation client/controller.

cloudformation.StackPolicyAssigner:

A StackPolicyAssigner is an Stack with an additional method for mapping
a controller provided role-name into it's template (required because of
the current implementation of Principals/Roles.

cloudformation.StackSecretOutputter

A StackSecretOutputter is a Stack with an additional method that can
name a kubernets Secret that will be created/updated with any
cloudformation template outputs.

With these interfaces it was possible to write a generic controller
reconciler for any type that implements those interfaces, and means that
we no longer need to have seperate boilerplate types for
specs/controllers/templates. We can now direcly implement Stack
interfaces on spec types (saving the mapping of fields) and can
instaniate cloudformation.Controller for a given type (saving the need
to write boilerplate controllers).

A more generic interface object.Service and object.Status (intented to
be implmented by more object types than just cloudformation-based ones)
allow the sharing of the Status fields on objects, further reducing the
boilerplate required when implementing a new service type.

With the controllers fully generic, it was possible to create an
interface version of the AWS client (sdk.Client) which exposes only the
AWS SDK methods we make use of. Enabling mocking the client for use in
tests.

Integration tests have been added that can be run either in a mocked
mode (AWS_INTEGRATION=false) or against a real AWS environment
(AWS_INTEGRATION=true). The tests spin up a real kubernetes control
plane, spin up a controller manager, add the controllers to the
controller manager (in a way close to how the manager really starts up)
and then attempts to perform a create/update/delete of the resources
under test.

Examples of running the integration test can be found in
hack/test_integration.sh