alwways update status even if reconcile errors
after running reconcile we attempt to update any changes we have made to
our resource object.

if the reconcile errors, we _still_ want to try the update, but
returning an err from the CreateOrUpdate mutate func will cause it to
bail out, so instead we should check the errors seperately to ensure
update is attempted.