Allow specifying workers to run on spot instances
This introduces a new cluster variable,
`worker_on_demand_percentage_above_min`, which defines the percentage
of instances that should be on-demand instances.  It defaults to 100.
But you can set it to (for example) 50, which means that the first
`extra_workers_per_az_count` instances in each AZ nodegroup will be
on-demand, then half the workers beyond that will be on-demand and
half will be spot.

As a result, I expect this change to do nothing on its own, but we
could experiment with setting `worker-on-demand-percentage-above-min`
to different values in tech-ops-cluster-config.

Spot instances require supplying a list of allowed instance types to
use.  I picked instances based on the regex `[cmr]5a?d.xlarge`.  My
criteria were:

 - we seem to like instance storage
 - we want instances to be at least `xlarge` so that we get 60 IPs per
   instance
 - we are, in our current workload, more likely to be IP-bound rather
   than either CPU- or memory-bound
 - we don't care if we run Intel or AMD

So I think any of these instance types fit those criteria.  In terms
of on-demand pricing, `c5ad` is cheapest and `r5d` is most expensive,
reflecting the fact that a) AMD is cheaper than Intel, and b) `c5` has
less memory than `m5` which has less than `r5`.