Write ADR for safe replacement of nodes
To try to improve platform reliability during platform change operations.

As an SRE I sometimes need to change AMI (for an EKS update), instance type, or
anything else in the launch template in some way (e.g. instance role), and
doing this results in the replacement of all nodes.
In addition sometimes this may be triggered by AWS releasing a new AMI for our
EKS version.

This breaks a lot right now and we should fix that.

Co-Authored-By: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>
Co-Authored-By: Chris Farms <chris.farmiloe@digital.cabinet-office.gov.uk>
Co-Authored-By: Daniel Blair <daniel.blair@digital.cabinet-office.gov.uk>
Co-Authored-By: Philip Potter <philip.potter@digital.cabinet-office.gov.uk>