svcop: adds ImageRepository (ECR) resources
Adds support for ImageRepository resources to the service-operator to
enable tenant provisioning of container image respositories.

Implements `cloudformation.Stack` for the provisioning of AWS ECR
repositories in default registry of the account that the service
operator is deployed into.

To provision an ECR repository create an `ImageRepository` resource and
a `Principal` that will gain the permisisons to access the repository.

The operator will provision a cloudformation stack to create an ECR
respository with the same name as the resource. Image name / URL will be
output to a `Secret` with the same name as the resource (or can be
overriden by setting the `spec.secret` property.

The target principal will have an IAM policy attached with full access
to the provisioned ecr repository and will also have
`ecr:GetAuthorizationToken` permission so that the principal's IAM role
can be used to generate temporary credentials suitable for use with
`docker push`.

In order to make consuming the `docker push` credentials more
user-friendly (ecr credneitals only last 12hrs), the Principal resource
will maintain a Secret containing ImageRegistryUsername and
ImageRegsitryPassword that is refreshed every 6hrs. This make it simple
to consume from concourse in a way that does not break namespace
isolation (ie, we can't use instance profiles on the shared workers).

To facilitate periodic generation of docker credentials:

* the service-operator's role needs to be able to assume the IAM roles
  that it provisions. (assume-role -> generate-creds-for-that-role),
  which has required modification of the sdk.Client interface,
  terraformed roles and policies of all existing Principal-roles.
* The requeue intervals/timeouts have been made more configurable so
  that we can requeue requests when no errors have occured ensuring
  secrets are updated within the 12hr validity.
* An additional interface / step to the reconciler has been added so
  that non-cloudformation based logic can contribute keys/values to the
  output secrets of `Stacks`

Test coverage for provisioning against both fakes/aws has been added.