[untested] work around concourse volume problem in deployer pipeline
I think we've hit a concourse bug when we upgraded from concourse
6.0.0 to 6.3.0.  I've filed concourse/concourse#5799 upstream to
capture this.  In the meantime, I think we can work around this by
forcing each task to use a different volume.

This fix works by avoiding reusing the same volume; instead, we create
new volumes at each step and copy the files from the old into the new.

Some details to note:

  - the old `platform` volume was somewhat abused - it was both the
    assets from the github release and the untarred source code from
    the git repo.  this commit splits those up into a
    `platform-release` volume and a `platform-source` volume.
  - in the `deploy` job, we modify the source multiple times, so
    we also need volumes called `platform-source-with-hook` and
    `platform-merged-source`
  - similarly, in the `destroy` job we were modifying the source
    several times.  I collapsed two tasks into one badly-named one
    because the faff of managing volumes was too painful
  - this probably should be tested with on-demand clusters too because
    it interacts nontrivially with PLATFORM_RESOURCE_TYPE

I don't know how this interacts with `put` steps (in particular, when
we `put: cluster-state`).  I'm hoping that the `terraform_source:`
field works with volumes the way that I expect?

Let's merge and see if it fixes things?  It can't break them more than
they are already...